refactor: 디렉토리명 수정(avgTemperature -> AverageTemperature) - 축약형 수정
